remove Zoomer to Layer
